fix: move task-error api spec to core-services
